Moved Router, DB, JWT outside api
