+++ server/router.go
@@ -2,14 +2,13 @@ package server
 
 import (
 	"github.com/gin-gonic/gin"
-	"github.com/mkrs2404/eKYC/api/controllers"
 )
 
 func InitializeRouter(serverPort string) {
 
 	//Setting up Router
 	router := gin.Default()
-	controllers.InitializeRoutes(router)
+	InitializeRoutes(router)
 
 	//Starting up Server
 	router.Run(serverPort)
